feat: useCareerSectionフックを追加し、経歴データの処理を実装
